+++ e2e/jest-css-transform-demo/src/Button.spec.tsx
@@ -1,0 +1,3 @@
+import Button from './Button';
+
+it('验证button组件存在', () => expect(Button).toBeDefined());