+++ e2e/jest-css-transform-demo/src/Button.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import './Button.css';
+
+/**
+ * 用于测试的按钮组件
+ */
+export default function Button() {
+  return <div className="button">按钮</div>;
+}