+++ e2e/jest-css-transform-demo/src/Draftjs.spec.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+import { Editor } from 'draft-js';
+import 'draft-js/dist/Draft.css';
+
+it('editor', () => expect(<Editor />).toBeDefined());